+++ src/lightly_train/_models/dinov3/dinov3_src/hub/backbones.py
@@ -632,3 +632,62 @@ def dinov3_convnext_large(
     if not pretrained:
         model.init_weights()
     return model
+
+
+def _dinov3_vit_test(
+    *,
+    pretrained: bool = False,
+    weights: Union[Weights, str] = Weights.LVD1689M,
+    in_chans: int = 3,
+    **kwargs,
+):
+    return _make_dinov3_vit(
+        img_size=32,
+        patch_size=2,
+        in_chans=in_chans,
+        pos_embed_rope_base=100,
+        pos_embed_rope_normalize_coords="separate",
+        pos_embed_rope_rescale_coords=2,
+        pos_embed_rope_dtype="fp32",
+        embed_dim=64,
+        depth=2,
+        num_heads=4,
+        ffn_ratio=4,
+        qkv_bias=True,
+        drop_path_rate=0.0,
+        layerscale_init=1.0e-05,
+        norm_layer="layernormbf16",
+        ffn_layer="mlp",
+        ffn_bias=True,
+        proj_bias=True,
+        n_storage_tokens=0,
+        mask_k_bias=False,
+        pretrained=pretrained,
+        weights=weights,
+        compact_arch_name="vits",
+        check_hash=False,
+        **kwargs,
+    )
+
+
+def _dinov3_convnext_test(
+    *,
+    pretrained: bool = False,
+    weights: Union[Weights, str] = Weights.LVD1689M,
+    in_chans: int = 3,
+    **kwargs,
+):
+    model = _make_dinov3_convnext(
+        in_chans=in_chans,
+        depths=[2, 2, 2, 2],
+        dims=[16, 16, 32, 32],
+        compact_arch_name="convnext_tiny",
+        drop_path_rate=0,
+        layer_scale_init_value=1e-6,
+        pretrained=pretrained,
+        weights=weights,
+        **kwargs,
+    )
+    if not pretrained:
+        model.init_weights()
+    return model